Invert ownership of creating COM callable wrappers
Introduces an `IManagedWraper` to implement on classes that provide the COM interface table for ComWrappers to use. Implementation uses this interface to recognize supported types.

Using a combination of non-generic base interface and a derived interface `IVTable` expresses COM callable wrapper support on the COM structs coming from CsWin32.

This change also implements CCWs for a number of Ole control interfaces.